+++ index.html
@@ -231,7 +231,7 @@
   </template>
 
   <!-- Внешний сценарий -->
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
