+++ js/random-int.js
@@ -1,0 +1,8 @@
+// Функция получения случайных целых чисел.
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export {getRandomInt};
