+++ js/string-check.js
@@ -1,0 +1,6 @@
+// Функция для проверки максимальной длины строки.
+function checkLength(string, maxLength) {
+  return (string.length <= maxLength);
+}
+
+export {checkLength};
